requirements.txt
amqp==2.4.2
anyjson==0.3.3
Babel==2.6.0
backports.csv==1.0.7
billiard==3.6.0.0
celery==4.3.0
certifi==2019.3.9
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.1.7
django-crispy-forms==1.7.2
django-formtools==2.1
django-import-export==1.2.0
django-pure-pagination==0.3.0
django-qiniu-storage==2.3.1
django-ranged-response==0.2.0
django-simple-captcha==0.5.11
DjangoUeditor==1.8.143
et-xmlfile==1.0.1
flower==0.9.3
future==0.17.1
httplib2==0.12.3
idna==2.8
jdcal==1.4.1
Jinja2==2.10.1
kombu==4.5.0
Markdown==3.1
MarkupSafe==1.1.1
meld3==1.0.2
numpy==1.16.3
odfpy==1.4.0
olefile==0.46
opencv-python==4.1.0.25
openpyxl==2.6.2
Pillow==6.0.0
PyMySQL==0.9.3
pytz==2019.1
PyYAML==5.1
qiniu==7.2.4
redis==3.2.1
requests==2.21.0
six==1.12.0
sqlparse==0.3.0
supervisor==4.0.2
tablib==0.13.0
tornado==5.1.1
urllib3==1.24.3
vine==1.3.0
xlrd==1.2.0
xlwt==1.3.0
